+++ src/main/kotlin/com/coder/gateway/models/CoderWorkspacesWizardModel.kt
@@ -4,6 +4,6 @@ data class CoderWorkspacesWizardModel(
     var coderURL: String = "https://localhost",
     var token: String = "",
     var buildVersion: String = "",
-    var workspaceAgents: List<WorkspaceAgentModel> = mutableListOf(),
+    var localCliPath: String = "",
     var selectedWorkspace: WorkspaceAgentModel? = null
 )