+++ src/main/kotlin/com/coder/gateway/sdk/ex/exceptions.kt
@@ -7,3 +7,5 @@ class AuthenticationResponseException(reason: String) : IOException(reason)
 class WorkspaceResponseException(reason: String) : IOException(reason)
 
 class WorkspaceResourcesResponseException(reason: String) : IOException(reason)
+
+class TemplateResponseException(reason: String) : IOException(reason)