+++ src/main/kotlin/com/coder/gateway/sdk/v2/CoderV2RestFacade.kt
@@ -1,11 +1,16 @@
 package com.coder.gateway.sdk.v2
 
 import com.coder.gateway.sdk.v2.models.BuildInfo
+import com.coder.gateway.sdk.v2.models.CreateWorkspaceBuildRequest
+import com.coder.gateway.sdk.v2.models.Template
 import com.coder.gateway.sdk.v2.models.User
 import com.coder.gateway.sdk.v2.models.Workspace
+import com.coder.gateway.sdk.v2.models.WorkspaceBuild
 import com.coder.gateway.sdk.v2.models.WorkspaceResource
 import retrofit2.Call
+import retrofit2.http.Body
 import retrofit2.http.GET
+import retrofit2.http.POST
 import retrofit2.http.Path
 import retrofit2.http.Query
 import java.util.UUID
@@ -27,6 +32,15 @@ interface CoderV2RestFacade {
     @GET("api/v2/buildinfo")
     fun buildInfo(): Call<BuildInfo>
 
-    @GET("api/v2/workspacebuilds/{buildID}/resources")
-    fun workspaceResourceByBuild(@Path("buildID") build: UUID): Call<List<WorkspaceResource>>
+    @GET("api/v2/templateversions/{templateID}/resources")
+    fun templateVersionResources(@Path("templateID") templateID: UUID): Call<List<WorkspaceResource>>
+
+    /**
+     * Queues a new build to occur for a workspace.
+     */
+    @POST("api/v2/workspaces/{workspaceID}/builds")
+    fun createWorkspaceBuild(@Path("workspaceID") workspaceID: UUID, @Body createWorkspaceBuildRequest: CreateWorkspaceBuildRequest): Call<WorkspaceBuild>
+
+    @GET("api/v2/templates/{templateID}")
+    fun template(@Path("templateID") templateID: UUID): Call<Template>
 }