+++ CHANGELOG.md
@@ -3,14 +3,28 @@
 # coder-gateway Changelog
 
 ## [Unreleased]
-
-## [2.0.2]
 ### Added
-- support for displaying working and non-working workspaces
-- better support for Light and Dark themes in the "Status" column 
+- support for displaying workspace version
+- support for managing the lifecycle of a workspace, i.e. start and stop and update workspace to the latest template version
 
+### Changed
+- workspace panel is now updated every 5 seconds
+- combinations of workspace names and agent names are now listed even when a workspace is down 
+- minimum supported Gateway build is now 222.3739.40
 
 ### Fixed
+- terminal link for workspaces with a single agent
+- no longer allow users to open a connection to a Windows or macOS workspace. It's not yet supported by Gateway
+
+## [2.0.2]
+
+### Added
+
+- support for displaying working and non-working workspaces
+- better support for Light and Dark themes in the "Status" column
+
+### Fixed
+
 - left panel is no longer visible when a new connection is triggered from Coder's "Recent Workspaces" panel.
   This provides consistency with other plugins compatible with Gateway
 - the "Select IDE and Project" button in the "Coder Workspaces" view is now disabled when no workspace is selected